Added new environment variable MCS_SAFE_MODE to allow the Mono build environment to switch off extensibility features while bootstrapping itself
Relying on the BOOTSTRAP_BASIC and BOOTSTRAP_NET_4_0 preprocessor defines is insufficient for determining when to load in add-ins
that consume the extensibility features or not. Middle phases of the Mono build, for example, build with only a partially
implemented version of System.dll which omits classes in the System.Configuration namespace. AddInManager relies on these classes to
load in the application configuration file and extract values from it. We thus check the value of the MCS_SAFE_MODE environment
variable ("true" means use safe mode) to determine if we should bypass add-in loading or not.